# java/conf/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "DB to REST",
        "description": "Shows the content of your database throw a REST interface",
        "version": "1.0"
    },
    "host": "localhost:8000/DBtoREST",
    "schemes": [
        "https",
        "http"
    ],
    "basePath": "/",
    "produces": [
        "application/json",
        "application/xml"
    ],
    "paths": {
        "/alldbs": {
            "get": {
                "summary": "Show the databases available, except those databases which belong to the DB system",
                "description": "Your database client may handle multiple databases. A priory, the content\nof the database might not be known and we need to know which databases do\nwe have in the system.\n",
                "tags": [
                    "Databases"
                ],
                "responses": {
                    "200": {
                        "description": "The list of available databases",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Databases"
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/alltables/{db}": {
            "get": {
                "summary": "Show the tables available in a database",
                "description": "A database can contain multiple tables. By indicating the database, we get the list of tables\navailable for such database\n",
                "parameters": [
                    {
                        "name": "db",
                        "in": "path",
                        "description": "Name of the database.",
                        "required": true,
                        "type": "string",
                        "format": "text"
                    }
                ],
                "tags": [
                    "Tables"
                ],
                "responses": {
                    "200": {
                        "description": "An array of price estimates by product",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Tables"
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/table/{db}/{table}": {
            "get": {
                "summary": "Shows the contents of a given table in the specified database",
                "description": "The main purpose of the application is to know exactly which are the contents of the database.\nTherefore, this is the main function of the project.\n",
                "parameters": [
                    {
                        "name": "db",
                        "in": "path",
                        "description": "Name of the database.",
                        "required": true,
                        "type": "string",
                        "format": "text"
                    },
                    {
                        "name": "table",
                        "in": "path",
                        "description": "Table in the specified database",
                        "required": true,
                        "type": "string",
                        "format": "text"
                    }
                ],
                "tags": [
                    "TableInfo"
                ],
                "responses": {
                    "200": {
                        "description": "Contents of the table",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/TableInfo"
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "Databases": {
            "type": "object"
        },
        "Tables": {
            "type": "object",
            "properties": {
                "database": {
                    "type": "string",
                    "description": "Existing database name"
                }
            }
        },
        "TableInfo": {
            "type": "object",
            "properties": {
                "database": {
                    "type": "string",
                    "description": "Name of the database"
                },
                "table": {
                    "type": "string",
                    "description": "Name of the table"
                }
            }
        },
        "Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "format": "int32"
                },
                "message": {
                    "type": "string"
                },
                "fields": {
                    "type": "string"
                }
            }
        }
    }
}